+++ src/index.tsx
@@ -1,2 +1,3 @@
+export * from './center-layout';
 export * from './column-layout';
 export * from './row-layout';