+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## Unreleased
 
+- CenterLayout: new component
 - ColumnLayout: new component
 - RowLayout: new component